this is new commit

+++ src/com/calculator/Calci.java
@@ -4,6 +4,6 @@ public interface Calci {
 	int add(int a,int b);
 	int sub(int a,int b);
 	double mul(int a,int b);
-	double div(int a, int b);
+	double div(double a, double b);
 
 }
